Add additional methods to HasCustomFields
